move platform() to testing-support
