--- libraries/apollo-runtime/src/appleTest/kotlin/test/network/utils.apple.kt
@@ -1,8 +1,0 @@
-package test.network
-
-
-/**
- * The current platform. This is used from tests because Double.toString() doesn't behave the same on JS and other platforms.
- * Prefer more specific functions like `assertMainThreadOnNative` when possible instead of checking the platform.
- */
-actual fun platform() = Platform.Native
