+++ libraries/apollo-testing-support/README.md
@@ -1,2 +1,7 @@
 # Module apollo-testing-support
 
+This module contains:
+
+* `QueueTestNetworkTransport` and `MapTestNetworkTransport` for testing without a mock server.
+* a set of helper functions used for Apollo tests. They were never really intended to become public and will be removed in a future version. These symbols are marked as deprecated. If you are using them, copy paste them in your project.
+
