+++ libraries/apollo-testing-support/src/jvmCommonMain/kotlin/com/apollographql/apollo3/testing/platform.jvmCommon.kt
@@ -1,0 +1,13 @@
+package com.apollographql.apollo3.testing
+
+import com.apollographql.apollo3.annotations.ApolloDeprecatedSince
+import com.apollographql.apollo3.annotations.ApolloInternal
+
+/**
+ * The current platform. This is used from tests because Double.toString() doesn't behave the same on JS and other platforms.
+ * Prefer more specific functions like `assertMainThreadOnNative` when possible instead of checking the platform.
+ */
+@ApolloInternal
+@ApolloDeprecatedSince(ApolloDeprecatedSince.Version.v4_0_0)
+@Deprecated("This function is not Apollo specific and will be removed in a future version. Copy/paste it in your codebase if you need it")
+actual fun platform() = Platform.Jvm
