+++ tests/defer/src/commonTest/kotlin/test/DeferNormalizedCacheTest.kt
@@ -17,16 +17,16 @@ import com.apollographql.apollo3.exception.ApolloException
 import com.apollographql.apollo3.exception.ApolloHttpException
 import com.apollographql.apollo3.exception.ApolloNetworkException
 import com.apollographql.apollo3.exception.CacheMissException
+import com.apollographql.apollo3.network.NetworkTransport
+import com.apollographql.apollo3.testing.Platform
+import com.apollographql.apollo3.testing.internal.runTest
+import com.apollographql.apollo3.testing.platform
 import com.apollographql.mockserver.MockServer
 import com.apollographql.mockserver.assertNoRequest
 import com.apollographql.mockserver.awaitRequest
 import com.apollographql.mockserver.enqueueError
 import com.apollographql.mockserver.enqueueMultipart
 import com.apollographql.mockserver.enqueueStrings
-import com.apollographql.apollo3.mpp.Platform
-import com.apollographql.apollo3.mpp.platform
-import com.apollographql.apollo3.network.NetworkTransport
-import com.apollographql.apollo3.testing.internal.runTest
 import com.benasher44.uuid.uuid4
 import defer.SimpleDeferQuery
 import defer.WithFragmentSpreadsMutation
