+++ tests/defer/src/commonTest/kotlin/test/DeferTest.kt
@@ -4,14 +4,14 @@ import com.apollographql.apollo3.ApolloClient
 import com.apollographql.apollo3.api.ApolloResponse
 import com.apollographql.apollo3.api.Error
 import com.apollographql.apollo3.autoPersistedQueryInfo
+import com.apollographql.apollo3.mpp.currentTimeMillis
+import com.apollographql.apollo3.testing.Platform
+import com.apollographql.apollo3.testing.internal.runTest
+import com.apollographql.apollo3.testing.platform
 import com.apollographql.mockserver.MockServer
 import com.apollographql.mockserver.enqueueMultipart
 import com.apollographql.mockserver.enqueueString
 import com.apollographql.mockserver.enqueueStrings
-import com.apollographql.apollo3.mpp.Platform
-import com.apollographql.apollo3.mpp.currentTimeMillis
-import com.apollographql.apollo3.mpp.platform
-import com.apollographql.apollo3.testing.internal.runTest
 import com.benasher44.uuid.uuid4
 import defer.SimpleDeferQuery
 import defer.WithFragmentSpreadsQuery
