+++ tests/engine/build.gradle.kts
@@ -29,6 +29,7 @@ kotlin {
       dependsOn(commonTest)
       dependencies {
         implementation(libs.apollo.engine.ktor)
+        implementation(libs.apollo.testingsupport)
         implementation(libs.slf4j.nop)
       }
     }
