+++ tests/multipart/src/commonTest/kotlin/test/MultipartSubscriptionsMockServerTest.kt
@@ -9,9 +9,11 @@ import com.apollographql.mockserver.MockServer
 import com.apollographql.apollo3.mpp.Platform
 import com.apollographql.apollo3.mpp.platform
 import com.apollographql.apollo3.network.http.HttpNetworkTransport
+import com.apollographql.apollo3.testing.Platform
 import com.apollographql.apollo3.testing.assertNoElement
 import com.apollographql.apollo3.testing.awaitElement
 import com.apollographql.apollo3.testing.internal.runTest
+import com.apollographql.apollo3.testing.platform
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.channels.Channel
